+++ native/src/main/java/io/ballerina/stdlib/http/transport/contractimpl/sender/states/http2/ReceivingEntityBody.java
@@ -18,6 +18,7 @@
 
 package io.ballerina.stdlib.http.transport.contractimpl.sender.states.http2;
 
+import io.ballerina.stdlib.http.api.logging.accesslog.SenderHttpAccessLogger;
 import io.ballerina.stdlib.http.transport.contractimpl.common.states.Http2MessageStateContext;
 import io.ballerina.stdlib.http.transport.contractimpl.sender.http2.Http2ClientChannel;
 import io.ballerina.stdlib.http.transport.contractimpl.sender.http2.Http2TargetHandler;
@@ -53,12 +54,17 @@ public class ReceivingEntityBody implements SenderState {
     private final Http2TargetHandler http2TargetHandler;
     private final Http2ClientChannel http2ClientChannel;
     private final Http2TargetHandler.Http2RequestWriter http2RequestWriter;
+    private SenderHttpAccessLogger accessLogger;
 
     ReceivingEntityBody(Http2TargetHandler http2TargetHandler,
                         Http2TargetHandler.Http2RequestWriter http2RequestWriter) {
         this.http2TargetHandler = http2TargetHandler;
         this.http2RequestWriter = http2RequestWriter;
         this.http2ClientChannel = http2TargetHandler.getHttp2ClientChannel();
+        if (http2TargetHandler.getHttpClientChannelInitializer().isHttpAccessLogEnabled()) {
+            accessLogger =
+                    new SenderHttpAccessLogger(http2TargetHandler.getHttp2ClientChannel().getChannel().remoteAddress());
+        }
     }
 
     @Override
@@ -135,6 +141,9 @@ public class ReceivingEntityBody implements SenderState {
                             Http2MessageStateContext http2MessageStateContext) {
         int streamId = http2DataFrame.getStreamId();
         ByteBuf data = http2DataFrame.getData();
+        if (accessLogger != null) {
+            accessLogger.updateContentLength(data);
+        }
         boolean endOfStream = http2DataFrame.isEndOfStream();
 
         if (serverPush) {
@@ -143,6 +152,9 @@ public class ReceivingEntityBody implements SenderState {
             onResponseDataRead(outboundMsgHolder, streamId, endOfStream, data);
         }
         if (endOfStream) {
+            if (accessLogger != null) {
+                accessLogger.updateAccessLogInfo(outboundMsgHolder.getRequest(), outboundMsgHolder.getResponse());
+            }
             http2MessageStateContext.setSenderState(new EntityBodyReceived(http2TargetHandler, http2RequestWriter));
         }
     }